requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aioresponses==0.7.8
aiosignal==1.3.2
altair==5.5.0
attrs==25.3.0
backoff==2.2.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
easyocr==1.7.1
filelock==3.18.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
imageio==2.37.0
inference-sdk==0.41.0
intel-openmp==2021.4.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.8.1
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.30.0
networkx==3.4.2
ninja==1.11.1.3
numpy==1.26.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==2.1.2
Pillow==10.1.0
propcache==0.3.0
protobuf==5.29.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyclipper==1.3.0.post6
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
scikit-image==0.25.2
scipy==1.15.2
shapely==2.0.7
six==1.17.0
smmap==5.0.2
st-cookies-manager==0.2.2
streamlit==1.36.0
supervision==0.25.1
sympy==1.13.3
tbb==2021.13.1
tenacity==8.5.0
tifffile==2025.3.13
toml==0.10.2
torch==2.3.1
torchvision==0.18.1
tornado==6.4.2
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==4.0.2
yarl==1.18.3
gradio_client